src/models/segmentation_model.py
"""
Segmentation Model Wrapper
Provides interface for food segmentation using Mask2Former or similar models
"""

from typing import List, Dict, Tuple, Optional
import numpy as np
from pathlib import Path
from loguru import logger

from ..utils import Config


class SegmentationModel:
    """Wrapper for food segmentation models"""
    
    def __init__(self, config: Optional[Config] = None):
        """
        Initialize segmentation model
        
        Args:
            config: Configuration object
        """
        self.config = config or Config()
        self.model_name = self.config.get("models.segmentation.name", "mask2former")
        self.device = self.config.get("models.segmentation.device", "cuda")
        self.confidence_threshold = self.config.get("models.segmentation.confidence_threshold", 0.5)
        
        self.model = None
        self.predictor = None
        
        logger.info(f"Initializing {self.model_name} segmentation model on {self.device}")
    
    def load_model(self):
        """
        Load the segmentation model
        
        This method should initialize:
        - Detectron2 Mask2Former model
        - Load pretrained weights
        - Configure predictor
        
        TODO: Implement model loading
        """
        config_file = self.config.get("models.segmentation.config")
        weights_path = self.config.get("models.segmentation.weights")
        
        logger.info(f"Loading model config: {config_file}")
        logger.info(f"Loading weights: {weights_path}")
        
        # TODO: Implement Detectron2 model loading
        # Example pseudocode:
        # from detectron2.config import get_cfg
        # from detectron2.engine import DefaultPredictor
        # 
        # cfg = get_cfg()
        # cfg.merge_from_file(config_file)
        # cfg.MODEL.WEIGHTS = weights_path
        # cfg.MODEL.DEVICE = self.device
        # self.predictor = DefaultPredictor(cfg)
        
        raise NotImplementedError("Segmentation model loading not yet implemented")
    
    def segment(self, image: np.ndarray) -> List[Dict]:
        """
        Perform semantic segmentation on an image
        
        Args:
            image: Input image as numpy array (H, W, 3)
        
        Returns:
            List of segmentation results, each containing:
                - mask: Binary mask (H, W)
                - bbox: Bounding box [x1, y1, x2, y2]
                - score: Confidence score
                - class_id: Predicted class ID
                - class_name: Class name
        
        TODO: Implement segmentation
        """
        if self.predictor is None:
            self.load_model()
        
        logger.debug(f"Running segmentation on image shape: {image.shape}")
        
        # TODO: Implement actual segmentation
        # outputs = self.predictor(image)
        # instances = outputs["instances"]
        # 
        # results = []
        # for i in range(len(instances)):
        #     if instances.scores[i] >= self.confidence_threshold:
        #         results.append({
        #             "mask": instances.pred_masks[i].cpu().numpy(),
        #             "bbox": instances.pred_boxes[i].tensor.cpu().numpy()[0],
        #             "score": float(instances.scores[i]),
        #             "class_id": int(instances.pred_classes[i]),
        #             "class_name": self._get_class_name(int(instances.pred_classes[i]))
        #         })
        
        raise NotImplementedError("Segmentation not yet implemented")
    
    def extract_crops(
        self, 
        image: np.ndarray, 
        segments: List[Dict],
        output_dir: Optional[Path] = None
    ) -> List[Tuple[np.ndarray, Path]]:
        """
        Extract cropped images for each segment
        
        Args:
            image: Original image
            segments: List of segmentation results
            output_dir: Directory to save crops (optional)
        
        Returns:
            List of (cropped_image, save_path) tuples
        """
        crops = []
        
        for i, segment in enumerate(segments):
            mask = segment["mask"]
            bbox = segment["bbox"]
            
            # Extract bounding box
            x1, y1, x2, y2 = map(int, bbox)
            cropped = image[y1:y2, x1:x2].copy()
            
            # Apply mask to crop
            mask_crop = mask[y1:y2, x1:x2]
            if len(cropped.shape) == 3:
                mask_crop = mask_crop[:, :, np.newaxis]
            cropped = cropped * mask_crop
            
            # Save if output directory provided
            if output_dir:
                output_dir = Path(output_dir)
                output_dir.mkdir(parents=True, exist_ok=True)
                save_path = output_dir / f"segment_{i}.png"
                
                from PIL import Image
                Image.fromarray(cropped.astype(np.uint8)).save(save_path)
            else:
                save_path = None
            
            crops.append((cropped, save_path))
        
        logger.debug(f"Extracted {len(crops)} crops from segments")
        return crops
    
    def _get_class_name(self, class_id: int) -> str:
        """
        Get class name from class ID
        
        TODO: Implement class name mapping
        """
        # TODO: Load class names from COCO or custom dataset
        return f"class_{class_id}"
    
    def visualize_segments(
        self, 
        image: np.ndarray,
        segments: List[Dict],
        output_path: Optional[Path] = None
    ) -> np.ndarray:
        """
        Visualize segmentation results on image
        
        Args:
            image: Original image
            segments: List of segmentation results
            output_path: Path to save visualization (optional)
        
        Returns:
            Image with overlaid segments
        """
        # TODO: Implement visualization with colored masks and labels
        logger.debug(f"Visualizing {len(segments)} segments")
        
        vis_image = image.copy()
        
        # TODO: Overlay masks with different colors
        # for segment in segments:
        #     mask = segment["mask"]
        #     color = self._get_random_color()
        #     vis_image[mask] = vis_image[mask] * 0.6 + color * 0.4
        
        if output_path:
            from PIL import Image
            Image.fromarray(vis_image.astype(np.uint8)).save(output_path)
        
        return vis_image
    
    def get_segment_stats(self, segments: List[Dict]) -> Dict:
        """
        Calculate statistics for segmentation results
        
        Args:
            segments: List of segmentation results
        
        Returns:
            Dictionary with statistics
        """
        if not segments:
            return {
                "num_segments": 0,
                "avg_confidence": 0.0,
                "total_area_pixels": 0
            }
        
        areas = [segment["mask"].sum() for segment in segments]
        confidences = [segment["score"] for segment in segments]
        
        return {
            "num_segments": len(segments),
            "avg_confidence": np.mean(confidences),
            "min_confidence": np.min(confidences),
            "max_confidence": np.max(confidences),
            "total_area_pixels": sum(areas),
            "avg_area_pixels": np.mean(areas),
            "min_area_pixels": np.min(areas),
            "max_area_pixels": np.max(areas)
        }
