requirements.txt
# Core dependencies
python>=3.9

# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0
detectron2>=0.6
opencv-python>=4.8.0
timm>=0.9.0

# Image Processing & Depth Estimation
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# LLM & Embeddings
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5

# Vector Database & Search
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support
lancedb>=0.3.0

# Data Processing
pandas>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# API & Web
requests>=2.31.0
aiohttp>=3.9.0

# Logging & Monitoring
loguru>=0.7.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Jupyter & Visualization
jupyter>=1.0.0
notebook>=7.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pydantic>=2.0.0
jsonschema>=4.19.0
